+++ admin/src/components/PaginationServerDriven.jsx
@@ -1,22 +1,41 @@
-import DoubleChevronRight from "../assets/icons/DoubleChevronRight";
-import ChevronRightPage from "../assets/icons/ChevronRightPage";
-import DoubleChevronLeft from "../assets/icons/DoubleChevronLeft";
-import ChevronLeftPage from "../assets/icons/ChevronLeftPage";
+import { HiChevronDoubleLeft } from "react-icons/hi";
+import { HiChevronDoubleRight } from "react-icons/hi";
+import { HiChevronLeft } from "react-icons/hi";
+import { HiChevronRight } from "react-icons/hi";
+
 /**
  * Ce composant va avec le HistoricServerDriven.
  * Il gère les données de pagination en provenance du serveur.
- * Et il reprend la forme de la pagination des reactlist.
  */
 import React from "react";
 
 const DEFAULT_DISPLAYED_PAGES = 3;
 
-export default function PaginationServerDriven({ pageCount, currentPage, count, itemsPerPage, itemsCount, className, changePage, displayedPages = DEFAULT_DISPLAYED_PAGES }) {
+export default function PaginationServerDriven({
+  pageCount,
+  currentPage,
+  count,
+  itemsPerPage,
+  itemsCount,
+  className,
+  changePage,
+  displayedPages = DEFAULT_DISPLAYED_PAGES,
+  size,
+  changeSize,
+}) {
   const lastDisplayPage = Math.min(pageCount - 1, Math.max(displayedPages, currentPage));
   const firstDisplayPage = Math.max(lastDisplayPage - displayedPages + 1, 1);
   const lastDisplayItem = currentPage * itemsPerPage + itemsCount;
-  const lastPage = Math.floor(count / 20);
-  const pages = [];
+  const lastPage = Math.floor(count / size);
+  let pages = [];
+  const totalhits = currentPage * itemsPerPage + itemsCount;
+  const actualHits = currentPage * itemsPerPage + 1;
+  const sizeOptions = [
+    { label: "10", value: 10 },
+    { label: "20", value: 20 },
+    { label: "40", value: 40 },
+    { label: "50", value: 50 },
+  ];
 
   switch (true) {
     case lastDisplayItem === count: //derniere page
@@ -24,13 +43,13 @@ export default function PaginationServerDriven({ pageCount, currentPage, count, 
         pages.push(<PageButton key={"page-" + i} page={i} changePage={changePage} active={currentPage === i} lastPage={lastPage} />);
       }
       break;
-    case lastDisplayItem >= count - 20: // avant derniere page
+    case lastDisplayItem >= count - size: // avant derniere page
       for (let i = firstDisplayPage; i <= lastDisplayPage; ++i) {
         pages.push(<PageButton key={"page-" + i} page={i} changePage={changePage} active={currentPage === i} lastPage={lastPage} />);
       }
       break;
 
-    case lastDisplayItem / 20 === 1 || lastDisplayItem / 20 === 2 || lastDisplayItem / 20 === 3: // page 1,2,3
+    case lastDisplayItem / size === 1 || lastDisplayItem / size === 2 || lastDisplayItem / size === 3: // page 1,2,3
       for (let i = firstDisplayPage; i <= lastDisplayPage; ++i) {
         pages.push(<PageButton key={"page-" + i} page={i} changePage={changePage} active={currentPage === i} lastPage={lastPage} />);
       }
@@ -43,6 +62,9 @@ export default function PaginationServerDriven({ pageCount, currentPage, count, 
       }
       break;
   }
+
+  pages = pages.filter((item) => item.props.page !== 0 && item.props.page !== lastPage); //on supprime la premiere et derniere page si elles y sont
+
   function goToPrevious(e) {
     e.preventDefault();
     if (currentPage > 0) {
@@ -67,33 +89,52 @@ export default function PaginationServerDriven({ pageCount, currentPage, count, 
   function goToNextX5(e) {
     e.preventDefault();
     if (lastDisplayItem < count && lastDisplayItem + 100 < count) {
-      changePage && changePage(currentPage + 5);
-    } else {
-      if (count % 20 === 0) {
-        changePage && changePage(lastPage - 1);
-      } else {
-        changePage && changePage(lastPage);
-      }
+      return changePage && changePage(currentPage + 5);
     }
+    if (count % size === 0) {
+      return changePage && changePage(lastPage - 1);
+    }
+    return changePage && changePage(lastPage);
   }
+
+  function checkSize(newSize) {
+    if (currentPage * newSize > count) {
+      let newLastPage = Math.floor(count / newSize);
+      changePage(newLastPage);
+    }
+    changeSize(newSize);
+  }
+
   return (
     <div className={`flex items-center justify-between gap-1 ${className}`}>
+      {changeSize ? (
+        <div className="text-xs flex gap-2 justify-center items-center text-[#242526]">
+          <select className="min-w-[56px] min-h-[32px] pl-2 border text-gray-600 rounded-md pb-1" value={size} onChange={(e) => checkSize(parseInt(e.target.value))}>
+            {sizeOptions.map((item) => (
+              <option key={item.label} value={item.value}>
+                {item.label}
+              </option>
+            ))}
+          </select>
+          Éléments par page
+        </div>
+      ) : null}
       <div className="text-[12px] text-[#242526] font-bold">
-        {currentPage * itemsPerPage + 1} <span className="font-normal">-</span> {currentPage * itemsPerPage + itemsCount} <span className="font-normal"> sur </span>{" "}
-        {count === 10000 ? "plus de 10000" : count}
+        {actualHits} <span className="font-normal">-</span> {totalhits} <span className="font-normal"> sur </span> {count === 10000 ? "plus de 10000" : count}
       </div>
       <div className="flex gap-1 items-center justify-center">
         <div className="flex justify-center items-center min-h-[32px] min-w-[65px] font-bold text-[12px] border border-gray-200 rounded-md border-solid">
           <button
-            href="#"
             onClick={goToPreviousX5}
-            className={`flex m-auto flex-none w-8 h-8 items-center justify-center border-r border-solid border-gray-200 ${
-              currentPage > 0 ? "cursor-pointer" : "cursor-not-allowed"
-            }`}>
-            <DoubleChevronLeft fill={currentPage > 0 ? "#6B7280" : "#E5E7EB"} />
+            className="flex m-auto flex-none w-8 h-8 items-center justify-center border-r border-solid border-gray-200"
+            style={currentPage > 0 ? { cursor: "pointer" } : { cursor: "not-allowed" }}>
+            <HiChevronDoubleLeft size={16} className={currentPage > 0 ? "text-gray-600" : "text-gray-200"} />
           </button>
-          <button href="#" onClick={goToPrevious} className={`flex flex-none w-8 m-auto items-center justify-center ${currentPage > 0 ? "cursor-pointer" : "cursor-not-allowed"}`}>
-            <ChevronLeftPage fill={currentPage > 0 ? "#6B7280" : "#E5E7EB"} />
+          <button
+            onClick={goToPrevious}
+            className="flex flex-none w-8 m-auto items-center justify-center"
+            style={currentPage > 0 ? { cursor: "pointer" } : { cursor: "not-allowed" }}>
+            <HiChevronLeft size={16} className={currentPage > 0 ? "text-gray-600" : "text-gray-200"} />
           </button>
         </div>
         <div className="flex justify-center items-center min-h-[32px] text-xs text-gray-600 border border-gray-200 rounded-md border-solid">
@@ -102,14 +143,14 @@ export default function PaginationServerDriven({ pageCount, currentPage, count, 
           {currentPage > 2 ? <div className="flex px-1 text-xs text-gray-400 border-gray-200 border-r border-solid min-h-[32px] items-center">...</div> : null}
 
           {pages}
-          {currentPage < (count % 20 === 0 ? lastPage - 3 : lastPage - 2) ? (
+          {currentPage < (count % size === 0 ? lastPage - 3 : lastPage - 2) ? (
             <div className="flex px-1 text-xs text-gray-400 border-gray-200 border-r border-solid min-h-[32px] items-center">...</div>
           ) : null}
           {lastPage !== 0 ? (
             <PageButton
-              page={count % 20 === 0 ? lastPage - 1 : lastPage}
+              page={count % size === 0 ? lastPage - 1 : lastPage}
               changePage={changePage}
-              active={currentPage === (count % 20 === 0 ? lastPage - 1 : lastPage)}
+              active={currentPage === (count % size === 0 ? lastPage - 1 : lastPage)}
               lastPage={lastPage}
               isLast={true}
             />
@@ -118,18 +159,16 @@ export default function PaginationServerDriven({ pageCount, currentPage, count, 
 
         <div className="flex justify-center items-center min-h-[32px] min-w-[65px] font-bold text-[12px] border border-gray-200 rounded-md border-solid">
           <button
-            href="#"
             onClick={goToNext}
-            className={`flex items-center justify-center flex-none w-8 h-8 m-auto border-r border-solid border-gray-200 ${
-              lastDisplayItem < count ? "cursor-pointer" : "cursor-not-allowed"
-            }`}>
-            <ChevronRightPage fill={lastDisplayItem < count ? "#6B7280" : "#E5E7EB"} />
+            className="flex items-center justify-center flex-none w-8 h-8 m-auto border-r border-solid border-gray-200"
+            style={lastDisplayItem < count ? { cursor: "pointer" } : { cursor: "not-allowed" }}>
+            <HiChevronRight size={16} className={lastDisplayItem < count ? "text-gray-600" : "text-gray-200"} />
           </button>
           <button
-            href="#"
             onClick={goToNextX5}
-            className={`flex items-center justify-center flex-none w-8 m-auto ${lastDisplayItem < count ? "cursor-pointer" : "cursor-not-allowed"}`}>
-            <DoubleChevronRight fill={lastDisplayItem < count ? "#6B7280" : "#E5E7EB"} />
+            className="flex items-center justify-center flex-none w-8 m-auto"
+            style={lastDisplayItem < count ? { cursor: "pointer" } : { cursor: "not-allowed" }}>
+            <HiChevronDoubleRight size={16} className={lastDisplayItem < count ? "text-gray-600" : "text-gray-200"} />
           </button>
         </div>
       </div>
@@ -148,7 +187,12 @@ function PageButton({ page, changePage, active, lastPage, isLast = false }) {
     return className;
   };
   return (
-    <button href="#" onClick={() => changePage(page)} className={`flex items-center justify-center flex-none w-8 h-8 m-auto ` + getClass()}>
+    <div
+      href="#"
+      onClick={() => changePage(page)}
+      className={`ml-[5px] flex min-h-[30px] cursor-pointer items-center justify-center rounded-[3px] border-[1px] border-[transparent] bg-[#f7fafc] py-[3px] px-[10px] text-[12px] text-[#242526] ${
+        active ? "font-bold" : "font-regular"
+      }`}>
       {page + 1}
     </button>
   );