+++ admin/src/scenes/centersV2/youngs/pointage.jsx
@@ -40,6 +40,7 @@ export default function Pointage({ updateFilter, isYoungCheckinOpen, focusedSess
   const [paramData, setParamData] = useState({
     page: 0,
   });
+  const [size, setSize] = useState(10);
 
   useEffect(() => {
     updateFilter(selectedFilters);
@@ -89,7 +90,7 @@ export default function Pointage({ updateFilter, isYoungCheckinOpen, focusedSess
     <div style={{ display: "flex", alignItems: "flex-start", width: "100%" }}>
       <div style={{ display: "flex", flexDirection: "column", flex: "1" }}>
         <div style={{ display: "flex", alignItems: "flex-start", width: "100%", height: "100%" }}>
-          <div className="relative flex-1">
+          <div className="relative flex-1 mb-4 rounded-lg">
             <div className="mx-4 flex flex-col">
               <div className="flex w-full flex-row justify-between">
                 <div className="flex items-center gap-2">
@@ -107,6 +108,7 @@ export default function Pointage({ updateFilter, isYoungCheckinOpen, focusedSess
                     setSelectedFilters={setSelectedFilters}
                     paramData={paramData}
                     setParamData={setParamData}
+                    size={size}
                   />
                   <div>
                     {youngSelected?.length > 0 ? (
@@ -320,6 +322,8 @@ export default function Pointage({ updateFilter, isYoungCheckinOpen, focusedSess
               paramData={paramData}
               setParamData={setParamData}
               currentEntryOnPage={data?.length}
+              size={size}
+              setSize={setSize}
               render={
                 <table className="mt-6 w-full">
                   <thead className="">