+++ api/src/controllers/elasticsearch/sessionphase1.js
@@ -11,22 +11,32 @@ const { buildNdJson, buildRequestBody, joiElasticSearch } = require("./utils");
 router.post("/:action(search|export)", passport.authenticate(["referent"], { session: false, failWithError: true }), async (req, res) => {
   try {
     // Configuration
+    const { user, body } = req;
     const searchFields = ["nameCentre", "cityCentre", "zipCentre", "codeCentre"];
-    const filterFields = ["department.keyword", "region.keyword", "cohort.keyword", "code.keyword", "placesLeft", "hasTimeSchedule.keyword", "typology.keyword", "domain.keyword"];
+    const filterFields = [
+      "department.keyword",
+      "region.keyword",
+      "cohort.keyword",
+      "code.keyword",
+      "placesLeft",
+      "hasTimeSchedule.keyword",
+      "typology.keyword",
+      "domain.keyword",
+      "hasPedagoProject.keyword",
+    ];
     const sortFields = [];
-
     // Authorization
     if (!canSearchInElasticSearch(req.user, "sessionphase1")) return res.status(403).send({ ok: false, code: ERRORS.OPERATION_UNAUTHORIZED });
 
     // Body params validation
-    const { queryFilters, page, sort, error, exportFields } = joiElasticSearch({ filterFields, sortFields, body: req.body });
+    const { queryFilters, page, sort, error, exportFields, size } = joiElasticSearch({ filterFields, sortFields, body: req.body });
     if (error) return res.status(400).send({ ok: false, code: ERRORS.INVALID_PARAMS });
 
     let contextFilters = [];
     if (req.user.role === ROLES.REFERENT_REGION) contextFilters.push({ term: { "region.keyword": req.user.region } });
     if (req.user.role === ROLES.REFERENT_DEPARTMENT) contextFilters.push({ terms: { "department.keyword": req.user.department } });
 
-    const { hitsRequestBody, aggsRequestBody } = buildRequestBody({ searchFields, filterFields, queryFilters, page, sort, contextFilters });
+    const { hitsRequestBody, aggsRequestBody } = buildRequestBody({ searchFields, filterFields, queryFilters, page, sort, contextFilters, size });
 
     if (req.params.action === "export") {
       const response = await allRecords("sessionphase1", hitsRequestBody.query, esClient, exportFields);