+++ app/src/scenes/auth/mobile/signin.jsx
@@ -17,6 +17,7 @@ import { environment } from "../../../config";
 import { Link } from "react-router-dom";
 import { isValidRedirectUrl } from "snu-lib/isValidRedirectUrl";
 import { captureMessage } from "../../../sentry";
+import plausibleEvent from "../../../services/plausible";
 
 export default function Signin() {
   const [email, setEmail] = React.useState("");
@@ -107,9 +108,9 @@ export default function Signin() {
         <div className="mt-4 text-[#E5E5E5]">
           <div className="mt-4 mb-3 text-center text-[17px] font-bold text-[#161616]">Vous n&apos;êtes pas encore inscrit(e) ?</div>
           <p className="text-center text-base text-[#161616] mb-4">
-            Les inscriptions sont actuellement uniquement ouvertes aux volontaires âgés de 15 et 17 ans et scolarisés en seconde GT en Nouvelle-Calédonie ou à Wallis-et-Futuna.
+            Les inscriptions sont actuellement uniquement ouvertes aux volontaires âgés de 15 à 17 ans et scolarisés en seconde en Nouvelle-Calédonie ou à Wallis-et-Futuna.
           </p>
-          <Link to="/preinscription">
+          <Link onClick={() => plausibleEvent("Connexion/Lien vers preinscription")} to="/preinscription">
             <p className="w-full my-4 text-center p-2 text-blue-france-sun-113 border-[1px] border-blue-france-sun-113 hover:bg-blue-france-sun-113">Pré-inscription</p>
           </Link>
           <br />