+++ app/src/scenes/preinscription/steps/stepNonEligible.jsx
@@ -53,7 +53,7 @@ export default function NonEligible() {
     <DSFRContainer>
       <h1 className="text-[22px] font-bold">Nous n'avons pas trouvé de séjour qui correspond à votre situation.</h1>
       <p className="mb-2 mt-4 border-l-8 border-l-[#6A6AF4] pl-4">
-        Les inscriptions sont actuellement uniquement ouvertes aux volontaires <strong>âgés de 15 et 17 ans</strong> et <strong>scolarisés en seconde GT</strong>{" "}
+        Les inscriptions sont actuellement uniquement ouvertes aux volontaires <strong>âgés de 15 à 17 ans</strong> et <strong>scolarisés en seconde</strong>{" "}
         <strong>en Nouvelle-Calédonie ou à Wallis-et-Futuna</strong>.
       </p>
       <p className="text-gray-500 mt-4">