+++ packages/lib/roles.js
@@ -555,7 +555,7 @@ function canViewStructureChildren(actor) {
 
 function canDownloadYoungDocuments(actor, target, type = null, _applications) {
   if (type === "certificate" || type === "convocation") {
-    return [ROLES.REFERENT_DEPARTMENT, ROLES.REFERENT_REGION].includes(actor.role);
+    return [ROLES.REFERENT_DEPARTMENT, ROLES.REFERENT_REGION, ROLES.ADMIN, ROLES.HEAD_CENTER, ROLES.RESPONSIBLE, ROLES.SUPERVISOR].includes(actor.role);
   } else {
     return (
       canEditYoung(actor, target) || [ROLES.RESPONSIBLE, ROLES.SUPERVISOR].includes(actor.role)